Relocating the remaining automower services.
