+++ src/automowerPlatform.ts
@@ -7,8 +7,8 @@ import { AutomowerAccessoryFactory, AutomowerAccessoryFactoryImpl } from './prim
 import { Localization, Y18nLocalization } from './primitives/localization';
 import { PlatformContainer, PlatformContainerImpl } from './primitives/platformContainer';
 import { AccessTokenManager, AccessTokenManagerImpl } from './services/husqvarna/accessTokenManager';
-import { DiscoveryService, DiscoveryServiceImpl } from './services/automower/discoveryService';
-import { EventStreamService, EventStreamServiceImpl } from './services/automower/eventStreamService';
+import { DiscoveryService, DiscoveryServiceImpl } from './services/husqvarna/automower/discoveryService';
+import { EventStreamService, EventStreamServiceImpl } from './services/husqvarna/automower/eventStreamService';
 import { PLATFORM_NAME, PLUGIN_ID } from './settings';
 
 /**
