+++ src/primitives/platformContainer.ts
@@ -12,10 +12,10 @@ import { HomebridgeImitationLogger } from '../diagnostics/platformLogger';
 import { ConsoleWrapperImpl } from '../diagnostics/primitives/consoleWrapper';
 import { DefaultErrorFactory } from '../errors/errorFactory';
 import { AccessTokenManagerImpl } from '../services/husqvarna/accessTokenManager';
-import { DiscoveryServiceImpl } from '../services/automower/discoveryService';
-import { EventStreamServiceImpl } from '../services/automower/eventStreamService';
-import { GetMowersServiceImpl } from '../services/automower/getMowersService';
-import { MowerControlServiceImpl } from '../services/automower/mowerControlService';
+import { DiscoveryServiceImpl } from '../services/husqvarna/automower/discoveryService';
+import { EventStreamServiceImpl } from '../services/husqvarna/automower/eventStreamService';
+import { GetMowersServiceImpl } from '../services/husqvarna/automower/getMowersService';
+import { MowerControlServiceImpl } from '../services/husqvarna/automower/mowerControlService';
 import { DeterministicMowerFaultedPolicy } from '../services/policies/mowerFaultedPolicy';
 import { DeterministicMowerInMotionPolicy } from '../services/policies/mowerInMotionPolicy';
 import { DeterministicMowerIsArrivingPolicy } from '../services/policies/mowerIsArrivingPolicy';
