+++ test/automowerPlatform.spec.ts
@@ -5,8 +5,8 @@ import { It, Mock, Times } from 'moq.ts';
 
 import { AutomowerPlatform, AutomowerPlatformConfig } from '../src/automowerPlatform';
 import { BadConfigurationError } from '../src/errors/badConfigurationError';
-import { DiscoveryService } from '../src/services/automower/discoveryService';
-import { EventStreamService } from '../src/services/automower/eventStreamService';
+import { DiscoveryService } from '../src/services/husqvarna/automower/discoveryService';
+import { EventStreamService } from '../src/services/husqvarna/automower/eventStreamService';
 import { PLATFORM_NAME, PLUGIN_ID } from '../src/settings';
 import { AutomowerPlatformSpy } from './automowerPlatformSpy';
