+++ test/services/husqvarna/automower/eventStreamServiceImpl.spec.ts
@@ -1,12 +1,12 @@
 import { It, Mock, Times } from 'moq.ts';
 
-import { PlatformLogger } from '../../../src/diagnostics/platformLogger';
-import { BadCredentialsError } from '../../../src/errors/badCredentialsError';
-import { AutomowerEventTypes, PositionsEvent, SettingsEvent, StatusEvent } from '../../../src/events';
-import { AccessToken, Activity, HeadlightMode, Mode, OverrideAction, RestrictedReason, State } from '../../../src/model';
-import { Timer } from '../../../src/primitives/timer';
-import { AccessTokenManager } from '../../../src/services/husqvarna/accessTokenManager';
-import { AutomowerEventStreamClientStub } from '../../clients/automowerEventStreamClientStub';
+import { PlatformLogger } from '../../../../src/diagnostics/platformLogger';
+import { BadCredentialsError } from '../../../../src/errors/badCredentialsError';
+import { AutomowerEventTypes, PositionsEvent, SettingsEvent, StatusEvent } from '../../../../src/events';
+import { AccessToken, Activity, HeadlightMode, Mode, OverrideAction, RestrictedReason, State } from '../../../../src/model';
+import { Timer } from '../../../../src/primitives/timer';
+import { AccessTokenManager } from '../../../../src/services/husqvarna/accessTokenManager';
+import { AutomowerEventStreamClientStub } from '../../../clients/automowerEventStreamClientStub';
 import { EventStreamServiceImplSpy } from './eventStreamServiceImplSpy';
 
 describe('EventStreamServiceImpl', () => {
