+++ test/services/scheduleSwitchImpl.spec.ts
@@ -6,7 +6,7 @@ import { AutomowerContext } from '../../src/automowerAccessory';
 import { PlatformLogger } from '../../src/diagnostics/platformLogger';
 import { Activity, Calendar, Mode, MowerState, Planner, RestrictedReason, State } from '../../src/model';
 import { NameMode } from '../../src/services/abstractSwitch';
-import { MowerControlService } from '../../src/services/automower/mowerControlService';
+import { MowerControlService } from '../../src/services/husqvarna/automower/mowerControlService';
 import { ScheduleEnabledPolicy } from '../../src/services/policies/scheduleEnabledPolicy';
 import { ScheduleSwitchImplSpy } from './scheduleSwitchImplSpy';
